fix(seo): route Organization JSON-LD through getOrigin() (#282)
